just changed return type from int to double for score in controller

+++ src/main/java/com/calmpainter/calm/painter/controller/GameController.java
@@ -40,7 +40,7 @@ public class GameController {
     }
     
     @GetMapping("/{gameId}/score")
-    public int calculateScore(@PathVariable String gameId) {
+    public double calculateScore(@PathVariable String gameId) {
         return gameService.calculateScore(gameId);
     }
